Move ipc handlers into its own module to fix specs
